Added ability to perform a refund by passing in the Id of the transaction that needs to be refunded.
Code added for refund response, and also the audit ID is now returned which is then saved on the Database as this is needed if attempting to perform a refund. The audit value does not need to be exposed back to the application layer, so this is not included in the payment response values that are sent back.
Some updates for other values that need to be sent through