Created a better way to process transactions when purchasing more than one event ticket, or when performing multiple donations to different funds in the same transaction. The system will will create the different transaction items as normal, and for a credit card transaction will simply combine the amounts of all transasction items and process as one, and then update each transaction item based on the result.
Direct debit transactions are queued as normal, and then I wrote some Linq in the ABA file generation to sum the amounts in the line items, and group by the bsb, account #, account name, and remitter (so that it can create an entirely new object of that type), and create the joined list, and then return it so we process it in the one transaction. Since this is only done in the ABA generation, it does not impact the transaction items, as they are already updated with awating clearance.